Select seats using file system waits
Only allocate fork #1 to the head of table when the head
is satisfied all other diners have taken their seats.

Other diners wait for fork #1 to turn up before progressing

This eliminates magic numbers in the simulation scripts and which
can now rely on the number of diners in the seating area not changing.