Switch to container structure
Setup now by running mkosi on the base directory. All the
extra files are in the mkosi.extra directory.
The 'nspawn' and 'mkosi.conf' files control the build.
Switch select-seat to inotify because fanotify now requires admin
privileges.
Update create philosopher to create the shared philosophers group
and create the default links in the home directory.